+++ packages/javascript/lib/name/index.ts
@@ -3,4 +3,3 @@ export { nameLengthRange } from './nameLengthRange.js';
 export { nameSupportsMiddleName } from './nameSupportsMiddleName.js';
 export { nameSupportsRoman } from './nameSupportsRoman.js';
 export { randName } from './randName.js';
-export { randNameDetails } from './randNameDetails.js';